Conclusion one
From this MatLab code, I decided to find the intersection of locations rated 4.1 or higher and what I deemed the locations with the highest tips- that is, where the minimum cost is $99 or above, or where the average cost is not a divisor of the minimum cost (such as a $40 average for a $50 minimum order). I then calculated the set of restaurants which meet both of these criteria, and plotted these points. When looking at the data (found in maxTips!.png), I found that, as a delivery person, staying in locations such as [39.5, -85.8] would be best for me to maximize my profits from tips because the higher rated restaurants are more likely to get traffic from users and higher priced orders have higher tips